changed prices back to bigint

+++ src/RoyalEstate.EntityFrameworkCore/Migrations/RoyalEstateDbContextModelSnapshot.cs
@@ -1665,8 +1665,8 @@ namespace RoyalEstate.Migrations
                     b.Property<long>("CustomerId")
                         .HasColumnType("bigint");
 
-                    b.Property<long?>("Deposit")
-                        .HasColumnType("bigint");
+                    b.Property<double?>("Deposit")
+                        .HasColumnType("float");
 
                     b.Property<string>("Description")
                         .HasColumnType("nvarchar(2000)")
@@ -1711,11 +1711,11 @@ namespace RoyalEstate.Migrations
                     b.Property<bool?>("Parking")
                         .HasColumnType("bit");
 
-                    b.Property<long?>("Price")
-                        .HasColumnType("bigint");
+                    b.Property<double?>("Price")
+                        .HasColumnType("float");
 
-                    b.Property<long?>("Rent")
-                        .HasColumnType("bigint");
+                    b.Property<double?>("Rent")
+                        .HasColumnType("float");
 
                     b.Property<int?>("Rooms")
                         .HasColumnType("int");
